requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.5
aiosignal==1.4.0
aiosqlite==0.20.0
annotated-types==0.7.0
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.7.14
frozenlist==1.7.0
idna==3.10
loguru==0.7.2
magic-filter==1.0.12
multidict==6.6.3
propcache==0.3.2
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
typing_extensions==4.14.1
yarl==1.20.1
